chore: change quote-line-item to vc-line-item
